requirements.txt
Django==1.6.2
Markdown==2.4
Pygments==1.6
SPARQLWrapper==1.5.2
argparse==1.2.1
django-filter==0.7
django-mptt==0.6.0
djangorestframework==2.3.12
html5lib==0.95
ipdb==0.8
ipython==1.2.0
isodate==0.4.9
lxml==3.3.0
pyparsing==2.0.1
rdflib==4.1.0
wsgiref==0.1.2
